requirements.txt
streamlit
pandas
numpy
folium
streamlit-folium
psycopg2-binary 